Update naming/docs for custom hooks
