--- hooks/useAuthOnClient.ts
@@ -1,33 +1,0 @@
-'use client'
-
-import { createClient } from '@/utils/supabase/client'
-import { User } from '@supabase/supabase-js'
-import { useEffect, useState } from 'react'
-
-// Client-side auth check
-export function useAuthOnClient() {
-  const [user, setUser] = useState<User | null>(null)
-  const [loading, setLoading] = useState(true)
-  const supabase = createClient()
-
-  useEffect(() => {
-    // Get initial session
-    supabase.auth.getUser().then(({ data: { user } }) => {
-      setUser(user)
-      setLoading(false)
-    })
-
-    // Listen for auth changes
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange((_event, session) => {
-      setUser(session?.user ?? null)
-    })
-
-    return () => {
-      subscription.unsubscribe()
-    }
-  }, [])
-
-  return { user, loading }
-}
